requirements.txt
streamlit
pandas
plotly.express 









0